added APIs and also implemented Express.Router() for organising API calls
